# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion utilisateur",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion utilisateur",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Déconnecté avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Infos de l'utilisateur connecté",
                "operationId": "8367feff40ba10e43187792cad3132c1",
                "responses": {
                    "200": {
                        "description": "Infos utilisateur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/breakdowns": {
            "get": {
                "tags": [
                    "Breakdown"
                ],
                "summary": "Liste des pannes de l'utilisateur connecté",
                "operationId": "8545c6d129e6bc37aa3b87a54c83f1d2",
                "responses": {
                    "200": {
                        "description": "Liste des pannes"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Breakdown"
                ],
                "summary": "Déclarer une nouvelle panne",
                "operationId": "7936319169c83fb434ee682d689f58f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_id",
                                    "description"
                                ],
                                "properties": {
                                    "vehicle_id": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Panne créée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/breakdowns/{id}": {
            "get": {
                "tags": [
                    "Breakdown"
                ],
                "summary": "Détail d'une panne",
                "operationId": "5967ae2386da9e21ec926853a6abac65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de la panne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/breakdowns/{id}/photos": {
            "post": {
                "tags": [
                    "Breakdown"
                ],
                "summary": "Ajouter une photo à une panne",
                "operationId": "fe75b08ff62911546a06c1488ea2c331",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "photo"
                                ],
                                "properties": {
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Photo ajoutée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Liste des notifications de l'utilisateur connecté",
                "operationId": "6d56a555955089a73fdf3f7229e8ead8",
                "responses": {
                    "200": {
                        "description": "Liste des notifications"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/{id}/read": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "Marquer une notification comme lue",
                "operationId": "89d94089af487b4bafdd6b9db9c55478",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marquée comme lue"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Liste des véhicules",
                "operationId": "043b3ce51f006f946511103cfe1c2ca1",
                "responses": {
                    "200": {
                        "description": "Liste des véhicules"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Détail d'un véhicule",
                "operationId": "874afdd20b81543018ea4033fc5e914c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du véhicule"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Breakdown",
            "description": "Breakdown"
        },
        {
            "name": "Notification",
            "description": "Notification"
        },
        {
            "name": "Vehicle",
            "description": "Vehicle"
        }
    ]
}